big rework, support sarc and yaz0
