+++ BfresToCast/Utils/SkeletonExtensions.cs
@@ -1,10 +1,10 @@
 ﻿using BfresLibrary;
 
-namespace BFResToCast;
+namespace BfresToCast.Utils;
 
 public static class SkeletonExtensions
 {
-    public static ushort GetBoneIndex(this Skeleton skeleton, float bone) => GetBoneIndex(skeleton, (int)bone);
+    public static ushort GetBoneIndex(this Skeleton skeleton, float bone) => skeleton.GetBoneIndex((int)bone);
     
     public static ushort GetBoneIndex(this Skeleton skeleton, int weightId)
     {
